LangChain_pdf_reader/requirements.txt
streamlit==1.35.0
python-dotenv==1.0.1
openai==1.33.0
langchain==0.2.3
langchain_community==0.2.4
pypdf==4.2.0
PyPDF2==3.0.1
sentence-transformers==3.0.1
chromadb==0.5.0